chore(dsp-api-error): refactor change TranferErrorToResponseTransformer to JsonObjectFromTransferErrorTransformer
